Updated Files

+++ backend/order/views.py
@@ -1,3 +1,4 @@
+import decimal
 from django.shortcuts import render
 from rest_framework.views import APIView
 # Create your views here.
@@ -17,15 +18,12 @@ class OrderPlaceAPIView(APIView):
 
     @transaction.atomic
     def post(self, request):
-
-        print(request.user.id)
         data = request.data
         customer = User.object.get(email=data['customer'])
-        print(f"DATA {customer}")
+    
 
         # Retrieve the cart for the user
         cart = Cart.objects.filter(customer=customer).first()
-        print(f"CART HAI {cart}")
         if not cart or not cart.items.exists():
             return Response({"detail": "Cart is empty or not found."}, status=status.HTTP_400_BAD_REQUEST)
 
@@ -46,15 +44,18 @@ class OrderPlaceAPIView(APIView):
 
         # Create order
         order_data = {
-            "customer": customer,
+            "customer": data['customer'],
             "total": total_price,
             "items": items,
         }
         serializer = OrderSerializer(data=order_data)
+        if not serializer.is_valid():
+            print("Validation Errors:", serializer.errors)  # Print errors
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
         serializer.is_valid(raise_exception=True)
         order = serializer.save()
 
         # Clear the cart
         cart.items.all().delete()
 
-        return Response(OrderSerializer(order).data, status=status.HTTP_201_CREATED)
+        return Response(OrderSerializer(order).data)
